(PC-26411)[PRO] refactor: remove RouteWrapper component
