+++ pro/src/app/AppRouter/AppRouter.tsx
@@ -8,25 +8,15 @@ import routes from 'app/AppRouter/routesMap'
 import NotFound from 'pages/Errors/NotFound/NotFound'
 import { selectActiveFeatures } from 'store/features/selectors'
 
-import { RouteWrapper } from './RouteWrapper'
-
 const sentryCreateBrowserRouter =
   Sentry.wrapCreateBrowserRouter(createBrowserRouter)
 
 const AppRouter = (): JSX.Element => {
   const activeFeatures = useSelector(selectActiveFeatures)
 
-  const activeRoutes = routes
-    .filter(
-      (route) =>
-        !route.featureName || activeFeatures.includes(route.featureName)
-    )
-    .map((route) => ({
-      ...route,
-      element: (
-        <RouteWrapper routeMeta={route.meta}>{route.element}</RouteWrapper>
-      ),
-    }))
+  const activeRoutes = routes.filter(
+    (route) => !route.featureName || activeFeatures.includes(route.featureName)
+  )
 
   const router = sentryCreateBrowserRouter([
     {
